+++ src/constants/form/coffee-chat-form.ts
@@ -1,0 +1,3 @@
+// 기존 null 이었던 introduction 값을 대체하기 위함
+// (지금은 해당 필드가 없으면 api 에러)
+export const INITIAL_INTRODUCTION = "커피챗 기본 소개글입니다"
