+++ src/interfaces/form.ts
@@ -69,21 +69,6 @@ export interface CommentUpdateFormData {
   commentForUpdate: string
 }
 
-// coffee chat
-export interface CoffeeChatFormData {
-  title: string
-  content: string
-  introduction: string
-  dateTimes?: string[]
-  hashTags?: string[]
-}
-
-export type CoffeeChatEditorInitialValues = CoffeeChatFormData
-
-export interface UpdateCoffeeChatData extends CoffeeChatFormData {
-  post_id: number
-}
-
 // hashtag
 type HashTag = string
