+++ src/page/coffee-chat/create/controls/rules/hashtag-rules.ts
@@ -1,6 +1,7 @@
 import { HASHTAG_LIMITS } from "@/constants/limitation"
 import { validationMessage } from "@/constants/message/validation"
-import { CoffeeChatFormData, HashTagFormData } from "@/interfaces/form"
+import { HashTagFormData } from "@/interfaces/form"
+import { CoffeeChatFormData } from "@/interfaces/form/coffee-chat-form"
 import { validatorInstance } from "@/util/validate"
 import { ControllerRenderProps, RegisterOptions } from "react-hook-form"
