+++ src/page/coffee-chat/create/controls/rules/introduction-rules.ts
@@ -1,5 +1,5 @@
 import { COFFEE_CHAT_LIMITS } from "@/constants/limitation"
-import { CoffeeChatFormData } from "@/interfaces/form"
+import { CoffeeChatFormData } from "@/interfaces/form/coffee-chat-form"
 import { RegisterOptions } from "react-hook-form"
 
 type ChatIntroductionRules = Omit<
